requirements.txt
--index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchaudio==2.5.1
espnet_model_zoo
librosa
numpy<2
scipy
soxr
transformers
accelerate
tqdm

fairseq

espnet==202412

# for calculating intrusive SE metrics
fastdtw
fast_bss_eval
diffsptk
pystoi

# for calculating the DNSMOS metric
onnx2torch
onnxruntime

# for calculating the phoneme similarity metric
levenshtein
phonemizer

# for text normalization in the WER metric
openai-whisper

# for SpeechBERTScore metric
# install a fork which supoprts multilingual SSL (MHuBERT-147)
# Removed pypesq dependency as we do not use it and it sometimes cause a problem when installing
https://github.com/ludlows/python-pesq/archive/master.zip
git+https://github.com/kohei0209/DiscreteSpeechMetrics.git@mhubert_wo_pypesq